fix: move to new package and module structure
